server/requirements.txt
openai-whisper
# Note: whisper requires PyTorch. Install torch appropriate for your platform.
# Example (CPU): pip install torch --index-url https://download.pytorch.org/whl/cpu
